+++ misskey-api/src/endpoint/roles.rs
@@ -1,3 +1,7 @@
 pub mod list;
 pub mod show;
 pub mod users;
+
+#[cfg(feature = "13-11-3")]
+#[cfg_attr(docsrs, doc(cfg(feature = "13-11-3")))]
+pub mod notes;